+++ Sources/PolygonioSwift/TickerSnapshot.swift
@@ -39,8 +39,8 @@ public struct TickerSnapshotResponse : Decodable {
         public var min: minuteOCHL
         public var prevDay: dayOCHL
         public var ticker: String
-        public var todaysChange: Decimal
-        public var todaysChangePerc: Decimal
+        public var todaysChange: Double
+        public var todaysChangePerc: Double
         public var updated: Int
         
         private enum CodingKeys: String, CodingKey {
@@ -58,11 +58,11 @@ public struct TickerSnapshotResponse : Decodable {
     
     public struct dayOCHL : Codable {
         public var volume: Int
-        public var VWAP: Decimal //VWAP (Volume Weighted Average Price)
-        public var open: Decimal
-        public var close: Decimal
-        public var high: Decimal
-        public var low: Decimal
+        public var VWAP: Double //VWAP (Volume Weighted Average Price)
+        public var open: Double
+        public var close: Double
+        public var high: Double
+        public var low: Double
 
         private enum CodingKeys: String, CodingKey {
             case volume = "v"
@@ -75,9 +75,9 @@ public struct TickerSnapshotResponse : Decodable {
     }
     
     public struct lastQuote : Codable {
-        public var askPrice: Decimal
+        public var askPrice: Double
         public var askSize: Int
-        public var bidPrice: Decimal
+        public var bidPrice: Double
         public var bidSize: Int
         public var timestamp: Int
 
@@ -93,7 +93,7 @@ public struct TickerSnapshotResponse : Decodable {
     public struct lastTrade : Codable {
         public var conditions: [Int]?
         public var tradeId: String
-        public var price: Decimal
+        public var price: Double
         public var size: Int
         public var timestamp: Int
         public var exchangeId: Int
@@ -111,11 +111,11 @@ public struct TickerSnapshotResponse : Decodable {
     public struct minuteOCHL : Codable {
         public var accumulatedVolume: Int
         public var volume: Int
-        public var VWAP: Decimal //VWAP (Volume Weighted Average Price)
-        public var open: Decimal
-        public var close: Decimal
-        public var high: Decimal
-        public var low: Decimal
+        public var VWAP: Double //VWAP (Volume Weighted Average Price)
+        public var open: Double
+        public var close: Double
+        public var high: Double
+        public var low: Double
 
         private enum CodingKeys: String, CodingKey {
             case accumulatedVolume = "av"
